requirements.txt
# Ego Vehicle Mask Generator Dependencies
# Python 3.10+

# Core
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0

# Configuration
PyYAML>=6.0

# GUI
PySide6>=6.5.0

# Utilities
tqdm>=4.65.0

